moved part of the code to separate new files, functions
