+++ src/components/ZipMarkdownToHtml/convertToHtml.js
@@ -1,0 +1,7 @@
+import showdown from 'showdown';
+
+export function convertToHtml(fileData) {
+  const converter = new showdown.Converter();
+  const html = converter.makeHtml(fileData);
+  return html;
+}
