+++ src/components/ZipMarkdownToHtml/createFile.js
@@ -1,0 +1,5 @@
+export function createFile(fileData, fileName) {
+  const blob = new Blob([fileData]);
+  const newFile = new File([blob], fileName);
+  return newFile;
+}
